Add CORE-V ALU tests and other minor adjustments
I'm adding tests for all the ALU instructions but:
  * exths, exthz, extbz, extbs
  * clip, clipr, clipu, clipur

In its current state, zero extension and sign extention
instructions are not behaving as expected. It currently gives
an excess error complaining about the input. The excess error
disappears when using the mode SI, however, it "extends" any value
to 0 - which is incorrect, so we end up with execution errors.

I've tried numerous different combinations of modes and tried
simplifying the code even further but to no avail.

One of the big changes regarding the already working instructions
are for min/max and minu/maxu.